+++ sdk/avs/arm-avs/samples/v7/javascript/workloadNetworksUpdatePortMirroringSample.js
@@ -8,7 +8,7 @@ const { DefaultAzureCredential } = require("@azure/identity");
  * This sample demonstrates how to update a WorkloadNetworkPortMirroring
  *
  * @summary update a WorkloadNetworkPortMirroring
- * x-ms-original-file: 2024-09-01/WorkloadNetworks_UpdatePortMirroring.json
+ * x-ms-original-file: 2025-09-01/WorkloadNetworks_UpdatePortMirroring.json
  */
 async function workloadNetworksUpdatePortMirroring() {
   const credential = new DefaultAzureCredential();