+++ sdk/avs/arm-avs/samples/v7/javascript/workloadNetworksUpdateSegmentsSample.js
@@ -8,7 +8,7 @@ const { DefaultAzureCredential } = require("@azure/identity");
  * This sample demonstrates how to update a WorkloadNetworkSegment
  *
  * @summary update a WorkloadNetworkSegment
- * x-ms-original-file: 2024-09-01/WorkloadNetworks_UpdateSegments.json
+ * x-ms-original-file: 2025-09-01/WorkloadNetworks_UpdateSegments.json
  */
 async function workloadNetworksUpdateSegments() {
   const credential = new DefaultAzureCredential();
@@ -17,10 +17,7 @@ async function workloadNetworksUpdateSegments() {
   const result = await client.workloadNetworks.updateSegments("group1", "cloud1", "segment1", {
     properties: {
       connectedGateway: "/infra/tier-1s/gateway",
-      subnet: {
-        dhcpRanges: ["40.20.0.0-40.20.0.1"],
-        gatewayAddress: "40.20.20.20/16",
-      },
+      subnet: { dhcpRanges: ["40.20.0.0-40.20.0.1"], gatewayAddress: "40.20.20.20/16" },
       revision: 1,
     },
   });