+++ sdk/avs/arm-avs/samples/v7/typescript/src/scriptCmdletsListSample.ts
@@ -8,7 +8,7 @@ import { DefaultAzureCredential } from "@azure/identity";
  * This sample demonstrates how to list ScriptCmdlet resources by ScriptPackage
  *
  * @summary list ScriptCmdlet resources by ScriptPackage
- * x-ms-original-file: 2024-09-01/ScriptCmdlets_List.json
+ * x-ms-original-file: 2025-09-01/ScriptCmdlets_List.json
  */
 async function scriptCmdletsList(): Promise<void> {
   const credential = new DefaultAzureCredential();