+++ sdk/compute/arm-compute/samples/v23/javascript/galleryInVMAccessControlProfileVersionsCreateOrUpdateSample.js
@@ -9,7 +9,7 @@ require("dotenv/config");
  * This sample demonstrates how to Create or update a gallery inVMAccessControlProfile version.
  *
  * @summary Create or update a gallery inVMAccessControlProfile version.
- * x-ms-original-file: specification/compute/resource-manager/Microsoft.Compute/GalleryRP/stable/2024-03-03/examples/galleryResourceProfileExamples/GalleryInVMAccessControlProfileVersion_Create.json
+ * x-ms-original-file: specification/compute/resource-manager/Microsoft.Compute/GalleryRP/stable/2025-03-03/examples/galleryResourceProfileExamples/GalleryInVMAccessControlProfileVersion_Create.json
  */
 async function createOrUpdateAGalleryInVMAccessControlProfileVersion() {
   const subscriptionId = process.env["COMPUTE_SUBSCRIPTION_ID"] || "{subscription-id}";