+++ sdk/containerregistry/arm-containerregistry/src/operationsInterfaces/tasks.ts
@@ -5,7 +5,6 @@
 // Changes may cause incorrect behavior and will be lost if the code is regenerated.
 
 import type { PagedAsyncIterableIterator } from "@azure/core-paging";
-import type { SimplePollerLike, OperationState } from "@azure/core-lro";
 import type {
   Task,
   TasksListOptionalParams,
@@ -13,10 +12,10 @@ import type {
   TasksGetResponse,
   TasksCreateOptionalParams,
   TasksCreateResponse,
-  TasksDeleteOptionalParams,
   TaskUpdateParameters,
   TasksUpdateOptionalParams,
   TasksUpdateResponse,
+  TasksDeleteOptionalParams,
   TasksGetDetailsOptionalParams,
   TasksGetDetailsResponse,
 } from "../models/index.js";
@@ -26,8 +25,8 @@ import type {
 export interface Tasks {
   /**
    * Lists all the tasks for a specified container registry.
-   * @param resourceGroupName The name of the resource group to which the container registry belongs.
-   * @param registryName The name of the container registry.
+   * @param resourceGroupName The name of the resource group. The name is case insensitive.
+   * @param registryName The name of the Registry
    * @param options The options parameters.
    */
   list(
@@ -37,8 +36,8 @@ export interface Tasks {
   ): PagedAsyncIterableIterator<Task>;
   /**
    * Get the properties of a specified task.
-   * @param resourceGroupName The name of the resource group to which the container registry belongs.
-   * @param registryName The name of the container registry.
+   * @param resourceGroupName The name of the resource group. The name is case insensitive.
+   * @param registryName The name of the Registry
    * @param taskName The name of the container registry task.
    * @param options The options parameters.
    */
@@ -50,28 +49,13 @@ export interface Tasks {
   ): Promise<TasksGetResponse>;
   /**
    * Creates a task for a container registry with the specified parameters.
-   * @param resourceGroupName The name of the resource group to which the container registry belongs.
-   * @param registryName The name of the container registry.
+   * @param resourceGroupName The name of the resource group. The name is case insensitive.
+   * @param registryName The name of the Registry
    * @param taskName The name of the container registry task.
    * @param taskCreateParameters The parameters for creating a task.
    * @param options The options parameters.
    */
-  beginCreate(
-    resourceGroupName: string,
-    registryName: string,
-    taskName: string,
-    taskCreateParameters: Task,
-    options?: TasksCreateOptionalParams,
-  ): Promise<SimplePollerLike<OperationState<TasksCreateResponse>, TasksCreateResponse>>;
-  /**
-   * Creates a task for a container registry with the specified parameters.
-   * @param resourceGroupName The name of the resource group to which the container registry belongs.
-   * @param registryName The name of the container registry.
-   * @param taskName The name of the container registry task.
-   * @param taskCreateParameters The parameters for creating a task.
-   * @param options The options parameters.
-   */
-  beginCreateAndWait(
+  create(
     resourceGroupName: string,
     registryName: string,
     taskName: string,
@@ -79,55 +63,14 @@ export interface Tasks {
     options?: TasksCreateOptionalParams,
   ): Promise<TasksCreateResponse>;
   /**
-   * Deletes a specified task.
-   * @param resourceGroupName The name of the resource group to which the container registry belongs.
-   * @param registryName The name of the container registry.
-   * @param taskName The name of the container registry task.
-   * @param options The options parameters.
-   */
-  beginDelete(
-    resourceGroupName: string,
-    registryName: string,
-    taskName: string,
-    options?: TasksDeleteOptionalParams,
-  ): Promise<SimplePollerLike<OperationState<void>, void>>;
-  /**
-   * Deletes a specified task.
-   * @param resourceGroupName The name of the resource group to which the container registry belongs.
-   * @param registryName The name of the container registry.
-   * @param taskName The name of the container registry task.
-   * @param options The options parameters.
-   */
-  beginDeleteAndWait(
-    resourceGroupName: string,
-    registryName: string,
-    taskName: string,
-    options?: TasksDeleteOptionalParams,
-  ): Promise<void>;
-  /**
    * Updates a task with the specified parameters.
-   * @param resourceGroupName The name of the resource group to which the container registry belongs.
-   * @param registryName The name of the container registry.
+   * @param resourceGroupName The name of the resource group. The name is case insensitive.
+   * @param registryName The name of the Registry
    * @param taskName The name of the container registry task.
    * @param taskUpdateParameters The parameters for updating a task.
    * @param options The options parameters.
    */
-  beginUpdate(
-    resourceGroupName: string,
-    registryName: string,
-    taskName: string,
-    taskUpdateParameters: TaskUpdateParameters,
-    options?: TasksUpdateOptionalParams,
-  ): Promise<SimplePollerLike<OperationState<TasksUpdateResponse>, TasksUpdateResponse>>;
-  /**
-   * Updates a task with the specified parameters.
-   * @param resourceGroupName The name of the resource group to which the container registry belongs.
-   * @param registryName The name of the container registry.
-   * @param taskName The name of the container registry task.
-   * @param taskUpdateParameters The parameters for updating a task.
-   * @param options The options parameters.
-   */
-  beginUpdateAndWait(
+  update(
     resourceGroupName: string,
     registryName: string,
     taskName: string,
@@ -135,9 +78,22 @@ export interface Tasks {
     options?: TasksUpdateOptionalParams,
   ): Promise<TasksUpdateResponse>;
   /**
+   * Deletes a specified task.
+   * @param resourceGroupName The name of the resource group. The name is case insensitive.
+   * @param registryName The name of the Registry
+   * @param taskName The name of the container registry task.
+   * @param options The options parameters.
+   */
+  delete(
+    resourceGroupName: string,
+    registryName: string,
+    taskName: string,
+    options?: TasksDeleteOptionalParams,
+  ): Promise<void>;
+  /**
    * Returns a task with extended information that includes all secrets.
-   * @param resourceGroupName The name of the resource group to which the container registry belongs.
-   * @param registryName The name of the container registry.
+   * @param resourceGroupName The name of the resource group. The name is case insensitive.
+   * @param registryName The name of the Registry
    * @param taskName The name of the container registry task.
    * @param options The options parameters.
    */