+++ sdk/monitor/monitor-opentelemetry/src/traces/handler.ts
@@ -50,8 +50,10 @@ export class TraceHandler {
     this._metricHandler = metricHandler;
     this._instrumentations = [];
     // Check sampler precedence
-    if (this._config.tracesPerSecond && this._config.tracesPerSecond >= 0) {
-      // If tracesPerSecond is set, use RateLimitedSampler
+    if (this._config.sampler) {
+      this._sampler = this._config.sampler;
+    } else if (this._config.tracesPerSecond && this._config.tracesPerSecond > 0) {
+      // If tracesPerSecond is set to a positive number, use RateLimitedSampler
       this._sampler = new RateLimitedSampler(this._config.tracesPerSecond);
     } else {
       // Otherwise, use PercentageSampler with samplingRatio
@@ -89,9 +91,11 @@ export class TraceHandler {
   /**
    * Shutdown handler
    */
-  // eslint-disable-next-line @typescript-eslint/require-await
   public async shutdown(): Promise<void> {
     this._azureFunctionsHook.shutdown();
+    await this._batchSpanProcessor.shutdown();
+    await this._azureSpanProcessor.shutdown();
+    await this._azureExporter.shutdown();
   }
 
   /**